Backend/Activity/requirements.txt
boto3
flask
python-dotenv
flask-cors
